requirements.txt
# flask==3.0.0
# flask-cors==4.0.0
# gunicorn==21.2.0
# supabase==2.9.0
# python-dotenv==1.0.0
# deepface==0.0.91
# ultralytics==8.1.0
# opencv-python-headless==4.8.1.78
# numpy==1.26.4
# scipy==1.11.4
# PyJWT==2.8.0
# bcrypt==4.1.2
# pillow==10.1.0
# tensorflow>=2.17.0
# tf-keras>=2.17.0
# websockets==14.1
# httpx==0.27.2


flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
supabase==2.9.0
python-dotenv==1.0.0
ultralytics==8.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.4
scipy==1.11.4
PyJWT==2.8.0
bcrypt==4.1.2
pillow==10.1.0
websockets==14.1
httpx==0.27.2
torch==2.1.0
torchvision==0.16.0
